Repackage classes to experimentpage; add an intermediate class between the services and the controllers (i.e. JSON serializers) to cache results
