+++ src/main/java/uk/ac/ebi/atlas/search/GeneSearchService.java
@@ -3,7 +3,7 @@ package uk.ac.ebi.atlas.search;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.springframework.stereotype.Component;
-import uk.ac.ebi.atlas.experimentpage.TSnePlotSettingsService;
+import uk.ac.ebi.atlas.experimentpage.tsneplot.TSnePlotSettingsService;
 
 import java.util.HashMap;
 import java.util.List;
@@ -14,6 +14,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
 import java.util.function.Function;
 
+import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static java.util.stream.Collectors.toMap;
 
 @Component
@@ -35,14 +36,14 @@ public class GeneSearchService {
     }
 
     // Returns inferred cell types and organism parts for each experiment accession
-    public Map<String, Map<String, List<String>>> getFacets(List<String> cellIds) {
+    public ImmutableMap<String, Map<String, List<String>>> getFacets(List<String> cellIds) {
         return geneSearchDao.getFacets(
                 cellIds,
                 "inferred_cell_type", "organism", "organism_part");
     }
 
     // Map<Gene ID, Map<Experiment accession, Map<K, Cluster ID>>>
-    public Map<String, Map<String, Map<Integer, List<Integer>>>> getMarkerGeneProfile(String... geneIds) {
+    public ImmutableMap<String, Map<String, Map<Integer, List<Integer>>>> getMarkerGeneProfile(String... geneIds) {
 
         return fetchInParallel(
                 ImmutableSet.copyOf(geneIds),
@@ -51,9 +52,9 @@ public class GeneSearchService {
                         geneId));
     }
 
-    private <T> Map<String, T> fetchInParallel(Set<String> geneIds, Function<String, T> geneIdInfoProvider) {
+    private <T> ImmutableMap<String, T> fetchInParallel(Set<String> geneIds, Function<String, T> geneIdInfoProvider) {
         // If this becomes a resource hog, consider having the pool as a member of the class and reuse it every time
-        ForkJoinPool forkJoinPool = new ForkJoinPool();
+        var forkJoinPool = new ForkJoinPool();
         try {
 
             return
@@ -62,8 +63,8 @@ public class GeneSearchService {
                                     .map(geneId -> ImmutableMap.of(geneId, geneIdInfoProvider.apply(geneId)))
                                     .map(Map::entrySet)
                                     .flatMap(Set::stream)
-                                    .collect(toMap(Map.Entry::getKey, Map.Entry::getValue))
-                    ).get();
+                                    .collect(toImmutableMap(Map.Entry::getKey, Map.Entry::getValue)))
+                            .get();
 
         } catch (InterruptedException | ExecutionException e) {
             throw new RuntimeException(e);
@@ -72,13 +73,13 @@ public class GeneSearchService {
         }
     }
 
-    private Map<String, Map<Integer, List<Integer>>> fetchClusterIDWithPreferredKAndMinPForGeneID(
+    private ImmutableMap<String, Map<Integer, List<Integer>>> fetchClusterIDWithPreferredKAndMinPForGeneID(
             List<String> experimentAccessions, String geneId) {
-        Map<String, Map<Integer, List<Integer>>> result = new HashMap<>();
+        var result = ImmutableMap.<String, Map<Integer, List<Integer>>>builder();
 
-        for (String experimentAccession : experimentAccessions) {
-            Optional<Integer> preferredK = tsnePlotSettingsService.getExpectedClusters(experimentAccession);
-            Map<Integer, List<Integer>> clusterIDWithPreferredKAndMinP =
+        for (var experimentAccession : experimentAccessions) {
+            var preferredK = tsnePlotSettingsService.getExpectedClusters(experimentAccession);
+            var clusterIDWithPreferredKAndMinP =
                     geneSearchDao.fetchClusterIdsWithPreferredKAndMinPForExperimentAccession(
                             geneId,
                             experimentAccession,
@@ -88,7 +89,6 @@ public class GeneSearchService {
                 result.put(experimentAccession, clusterIDWithPreferredKAndMinP);
             }
         }
-        return result;
-
+        return result.build();
     }
 }
