--- src/test/java/uk/ac/ebi/atlas/metadata/JsonCellMetadataControllerWIT.java
@@ -1,119 +1,0 @@
-package uk.ac.ebi.atlas.metadata;
-
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.http.MediaType;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import uk.ac.ebi.atlas.configuration.TestConfig;
-import uk.ac.ebi.atlas.testutils.JdbcUtils;
-
-import javax.inject.Inject;
-import javax.sql.DataSource;
-
-import java.util.stream.Collectors;
-
-import static org.hamcrest.Matchers.greaterThanOrEqualTo;
-import static org.hamcrest.Matchers.hasSize;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@ExtendWith(SpringExtension.class)
-@WebAppConfiguration
-@ContextConfiguration(classes = TestConfig.class)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class JsonCellMetadataControllerWIT {
-    @Inject
-    private DataSource dataSource;
-
-    @Inject
-    private JdbcUtils jdbcTestUtils;
-
-    @Autowired
-    private WebApplicationContext wac;
-
-    private MockMvc mockMvc;
-
-    @BeforeAll
-    void populateDatabaseTables() {
-        ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
-        populator.addScripts(
-                new ClassPathResource("fixtures/scxa_experiment-fixture.sql"),
-                new ClassPathResource("fixtures/scxa_analytics-fixture.sql"));
-        populator.execute(dataSource);
-    }
-
-    @AfterAll
-    void cleanDatabaseTables() {
-        ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
-        populator.addScripts(
-                new ClassPathResource("fixtures/scxa_experiment-delete.sql"),
-                new ClassPathResource("fixtures/scxa_analytics-delete.sql"));
-        populator.execute(dataSource);
-    }
-
-    @BeforeEach
-    void setUp() {
-        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
-    }
-
-    @ParameterizedTest
-    @MethodSource("experimentsWithMetadataProvider")
-    void validJsonForExistingCellId(String experimentAccession) throws Exception {
-        String cellId = jdbcTestUtils.fetchRandomCellFromExperiment(experimentAccession);
-
-        this.mockMvc
-                .perform(get(
-                        "/json/experiment/" + experimentAccession + "/cell/" + cellId + "/metadata"))
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8))
-                .andExpect(jsonPath("$", hasSize(greaterThanOrEqualTo(0))));
-    }
-
-    @Test
-    void emptyJsonForInvalidCellId() throws Exception {
-        String experimentAccession = jdbcTestUtils.fetchRandomSingleCellExperimentAccession();
-        String cellId = "FOO";
-
-        this.mockMvc
-                .perform(get(
-                        "/json/experiment/" + experimentAccession + "/cell/" + cellId + "/metadata"))
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8))
-                .andExpect(jsonPath("$", hasSize(0)));
-    }
-
-    @Test
-    void validJsonForInvalidExperimentAccession() throws Exception {
-        String experimentAccession = "FOO";
-        String cellId = "BAR";
-
-        this.mockMvc
-                .perform(get("/json/experiment/" + experimentAccession + "/cell/" + cellId + "/metadata"))
-                .andExpect(status().is4xxClientError());
-    }
-
-    private Iterable<String> experimentsWithMetadataProvider() {
-        // E-GEOD-99058 does not have any metadata (factors or inferred cell types)
-        return jdbcTestUtils.fetchPublicSingleCellExperimentAccessions()
-                .stream()
-                .filter(accession -> !accession.equalsIgnoreCase("E-GEOD-99058"))
-                .collect(Collectors.toSet());
-    }
-}
-
